backend/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
opencv-python
torch
torchvision
rembg
python-multipart
